refactor(rt): extract vehicle matching helpers
